+++ marko.json
@@ -633,7 +633,7 @@
     "transformer": "./src/common/transformers/attribute-tags/index.js"
   },
   "<ebay-select>": {
-    "renderer": "./src/components/ebay-select/index.js",
+    "renderer": "./src/components/ebay-select/index.marko",
     "@*": {
       "targetProperty": null,
       "type": "expression"
