+++ src/components/ebay-select/index.marko
@@ -1,0 +1,69 @@
+import processHtmlAttributes from "../../common/html-attributes"
+
+static {
+    var ignoredAttributes = [
+        "class",
+        "style",
+        "grow",
+        "borderless",
+        "disabled",
+        "options"
+    ];
+}
+
+static {
+    var itemIgnoredAttributes = [
+        "class",
+        "style",
+        "value",
+        "text",
+        "selected",
+        "optgroup"
+    ];
+}
+
+$ var options = state.options;
+$ var selectedOption = options[state.selectedIndex];
+$ var renderedOptGroups = [];
+<span
+    class=[input.class, "select", input.borderless && "select--borderless"]
+    style=input.style>
+    <select
+        ...processHtmlAttributes(input, ignoredAttributes)
+        onChange("handleChange")
+        disabled=state.disabled>
+        <for|option| of=options>
+            $ {
+                var withinOptgroup =
+                                option.optgroup &&
+                                renderedOptGroups.indexOf(option.optgroup) === -1;
+            }
+            <${option.optgroup ? null : "option"}
+                id:scoped="options"
+                key="options[]"
+                value=option.value
+                class=option.class
+                style=option.style
+                selected=(option === selectedOption)
+                ...processHtmlAttributes(option, itemIgnoredAttributes)>
+                <if(withinOptgroup)>
+                    <optgroup label=option.optgroup>
+                        <for|subOption| of=state.optGroups[option.optgroup]>
+                            <option
+                                value=subOption.value
+                                class=subOption.class
+                                style=subOption.style
+                                selected=(subOption === selectedOption)
+                                key="option[]">
+                                ${subOption.text}
+                            </option>
+                        </for>
+                    </optgroup>
+                    <% renderedOptGroups.push(option.optgroup); %>
+                </if>
+                <else-if(!option.optgroup)>${option.text}</else-if>
+            </>
+        </for>
+    </select>
+    <span class="select__icon"/>
+</span>
