--- src/components/ebay-select/template.marko
@@ -1,55 +1,0 @@
-<script marko-init>
-    var processHtmlAttributes = require("../../common/html-attributes");
-    var ignoredAttributes = ["class", "style", "grow", "borderless", "disabled", "options"];
-    var itemIgnoredAttributes = ["class", "style", "value", "text", "selected", "optgroup"]
-</script>
-
-<var options=data.options/>
-<var selectedOption=options[data.selectedIndex]/>
-<var renderedOptGroups=[]/>
-
-<span
-    w-bind
-    class=[
-        data.class,
-        "select",
-        data.borderless && "select--borderless"
-    ]
-    style=data.style>
-    <select
-        w-onchange="handleChange"
-        disabled=data.disabled
-        ${processHtmlAttributes(data, ignoredAttributes)}>
-        <for(option in options)>
-            <var withinOptgroup = (option.optgroup && renderedOptGroups.indexOf(option.optgroup) === -1)/>
-            <option
-                w-id="option[]"
-                value=option.value
-                class=option.class
-                style=option.style
-                selected=(option === selectedOption)
-                body-only-if(option.optgroup)
-                ${processHtmlAttributes(option, itemIgnoredAttributes)}>
-                <if(withinOptgroup)>
-                    <optgroup label=option.optgroup>
-                        <for(subOption in data.optGroups[option.optgroup])>
-                            <option
-                                w-id="option[]"
-                                value=subOption.value
-                                class=subOption.class
-                                style=subOption.style
-                                selected=(subOption === selectedOption)>
-                                ${subOption.text}
-                            </option>
-                        </for>
-                    </optgroup>
-                    <% renderedOptGroups.push(option.optgroup); %>
-                </if>
-                <else-if(!option.optgroup)>
-                    ${option.text}
-                </else-if>
-            </option>
-        </for>
-    </select>
-    <span class="select__icon"/>
-</span>
